requirements.txt
aws-cdk-lib>=2.26
aws-cdk.aws-glue-alpha>=2.25.0a0
aws-cdk.aws-redshift-alpha>=2.25.0a0
constructs>=10.0.0,<11.0.0

boto3>=1.23
pre-commit>=2.19
pylightxl>=1.59
pytest==6.2.5
sqlparse>=0.4.2
psycopg2>=2.9.3
wheel>=0.37.1